make on_(local_)main_process member of Accelerator
